+++ include/memory.h
@@ -11,8 +11,8 @@ extern char Memory[262145];
 extern int stackPointer;
 extern int StoreDataAddr;
 void ShowMemory(std::string address, int count);
-void storeMemory(int address, int Numbytes, int value);
-int fechMemory(int address, int Numbytes);
+void storeMemory(int address, int Numbytes, __int64 value);
+__int64 fechMemory(int address, int Numbytes);
 void showCallStack(int numLines);
 void storeData(std::vector<std::string> StoreData);
 #endif
